1.18 dev test

+++ src/main/java/net/bati/guilib/gui/screen/ScreenUtils.java
@@ -10,8 +10,7 @@ import java.util.Optional;
 
 
 public class ScreenUtils {
-
-    public static void renderWidgets(HashMap<String, Widget> widgets, MatrixStack matrices, int x, int y, float delta) {
+    public static void renderWidgets(HashMap<String, Widget> widgets, MatrixStack matrices, float x, float y, float delta) {
         Optional<Map.Entry<String, Widget>> widgetEntry = widgets.entrySet().stream().filter((entry) -> entry.getValue().isVisible() && entry.getValue().isHovered(x,y)).max(Comparator.comparingInt(current -> current.getValue().getRecursiveZ()));
 
         widgets.forEach((key, value) -> {
@@ -21,6 +20,7 @@ public class ScreenUtils {
                 value.setFocused(widgetEntry.isPresent() && key.contentEquals(widgetEntry.get().getKey()));
 
             value.render(matrices, x, y, delta);
+
         });
     }
 }
